Added ListingScope and refactored SortingModule to use it
