+++ app/src/main/java/com/esoxjem/movieguide/AppComponent.java
@@ -6,8 +6,6 @@ import com.esoxjem.movieguide.favorites.FavoritesModule;
 import com.esoxjem.movieguide.listing.ListingComponent;
 import com.esoxjem.movieguide.listing.ListingModule;
 import com.esoxjem.movieguide.network.NetworkModule;
-import com.esoxjem.movieguide.sorting.SortingDialogFragment;
-import com.esoxjem.movieguide.sorting.SortingModule;
 
 import javax.inject.Singleton;
 
@@ -21,12 +19,9 @@ import dagger.Component;
 @Component(modules = {
         AppModule.class,
         NetworkModule.class,
-        FavoritesModule.class,
-        SortingModule.class})
+        FavoritesModule.class})
 public interface AppComponent
 {
-    SortingDialogFragment inject(SortingDialogFragment fragment);
-
     DetailsComponent plus(DetailsModule detailsModule);
 
     ListingComponent plus(ListingModule listingModule);
