+++ app/src/main/java/com/esoxjem/movieguide/listing/ListingComponent.java
@@ -1,13 +1,18 @@
 package com.esoxjem.movieguide.listing;
 
+import com.esoxjem.movieguide.listing.sorting.SortingDialogFragment;
+import com.esoxjem.movieguide.listing.sorting.SortingModule;
+
 import dagger.Subcomponent;
 
 /**
  * @author arunsasidharan
  */
 @ListingScope
-@Subcomponent(modules = {ListingModule.class})
+@Subcomponent(modules = {ListingModule.class, SortingModule.class})
 public interface ListingComponent
 {
     MoviesListingFragment inject(MoviesListingFragment fragment);
+
+    SortingDialogFragment inject(SortingDialogFragment fragment);
 }
