+++ app/src/main/java/com/esoxjem/movieguide/listing/MoviesListingFragment.java
@@ -17,7 +17,7 @@ import android.widget.Toast;
 import com.esoxjem.movieguide.BaseApplication;
 import com.esoxjem.movieguide.Movie;
 import com.esoxjem.movieguide.R;
-import com.esoxjem.movieguide.sorting.SortingDialogFragment;
+import com.esoxjem.movieguide.listing.sorting.SortingDialogFragment;
 
 import java.util.ArrayList;
 import java.util.List;
